Move constraint-adapter tests into test_deprecations
These tests exercise deprecations.pre_process_constraints, which only
exists to support the deprecated legacy dict constraint API. Keeping
them in a dedicated test_constraint_adapter.py separated them from the
rest of the deprecation-era tests they should be deleted alongside once
dict constraints are removed in 0.6.0.
